+++ src/pipecat/services/google/google.py
@@ -601,23 +601,18 @@ class GoogleAssistantContextAggregator(OpenAIAssistantContextAggregator):
 
     async def handle_function_call_result(self, frame: FunctionCallResultFrame):
         if frame.result:
-            if not isinstance(frame.result, str):
-                return
-
-            response = {"response": frame.result}
-
+            await self._update_function_call_result(
+                frame.function_name, frame.tool_call_id, frame.result
+            )
+        else:
+            response = {"response": "COMPLETED"}
             await self._update_function_call_result(
                 frame.function_name, frame.tool_call_id, response
             )
-        else:
-            await self._update_function_call_result(
-                frame.function_name, frame.tool_call_id, "COMPLETED"
-            )
 
     async def handle_function_call_cancel(self, frame: FunctionCallCancelFrame):
-        await self._update_function_call_result(
-            frame.function_name, frame.tool_call_id, "CANCELLED"
-        )
+        response = {"response": "CANCELLED"}
+        await self._update_function_call_result(frame.function_name, frame.tool_call_id, response)
 
     async def _update_function_call_result(
         self, function_name: str, tool_call_id: str, result: Any
@@ -626,11 +621,12 @@ class GoogleAssistantContextAggregator(OpenAIAssistantContextAggregator):
             if message.role == "user":
                 for part in message.parts:
                     if part.function_response and part.function_response.id == tool_call_id:
-                        part.function_response.response = {"response": result}
+                        part.function_response.response = result
 
     async def handle_user_image_frame(self, frame: UserImageRawFrame):
+        response = {"response": "COMPLETED"}
         await self._update_function_call_result(
-            frame.request.function_name, frame.request.tool_call_id, "COMPLETED"
+            frame.request.function_name, frame.request.tool_call_id, response
         )
         self._context.add_image_frame_message(
             format=frame.format,